refactor: modularize core and simulation utility layers
